refactor(Any*): fix formatting

+++ test/strategies/any/AnyCompounderNaive.t.sol
@@ -9,24 +9,23 @@ import {AnyBaseTest} from "./AnyBase.t.sol";
 import "forge-std/console.sol";
 
 contract AnyCompounderNaiveImpl is AnyCompounderNaive {
-    function initialize(
-        address asset_,
-        address owner_,
-        bool autoDeposit_,
-        bytes memory strategyInitData_
-    ) external initializer {
+    function initialize(address asset_, address owner_, bool autoDeposit_, bytes memory strategyInitData_)
+        external
+        initializer
+    {
         __AnyConverter_init(asset_, owner_, autoDeposit_, strategyInitData_);
     }
 }
 
 contract ClaimContract {
     address[] public rewardTokens;
+
     constructor(address[] memory _rewardTokens) {
         rewardTokens = _rewardTokens;
     }
 
     fallback() external {
-        for (uint i; i < rewardTokens.length; i++) {
+        for (uint256 i; i < rewardTokens.length; i++) {
             IERC20(rewardTokens[i]).transfer(msg.sender, IERC20(rewardTokens[i]).balanceOf(address(this)));
         }
     }
@@ -36,29 +35,21 @@ contract AnyCompounderNaiveTest is AnyBaseTest {
     using stdJson for string;
 
     function setUp() public {
-        _setUpBaseTest(
-            0,
-            "./test/strategies/any/AnyCompounderNaiveTestConfig.json"
-        );
+        _setUpBaseTest(0, "./test/strategies/any/AnyCompounderNaiveTestConfig.json");
     }
 
-    function _setUpStrategy(
-        string memory json_,
-        string memory index_,
-        TestConfig memory testConfig_
-    ) internal override returns (IBaseStrategy) {
+    function _setUpStrategy(string memory json_, string memory index_, TestConfig memory testConfig_)
+        internal
+        override
+        returns (IBaseStrategy)
+    {
         AnyCompounderNaiveImpl _strategy = new AnyCompounderNaiveImpl();
         oracle = new MockOracle();
 
-        yieldAsset = json_.readAddress(
-            string.concat(".configs[", index_, "].specific.yieldAsset")
-        );
+        yieldAsset = json_.readAddress(string.concat(".configs[", index_, "].specific.yieldAsset"));
 
         _strategy.initialize(
-            testConfig_.asset,
-            address(this),
-            true,
-            abi.encode(yieldAsset, address(oracle), uint256(10), uint256(0))
+            testConfig_.asset, address(this), true, abi.encode(yieldAsset, address(oracle), uint256(10), uint256(0))
         );
 
         _strategy.setRewardTokens(json.readAddressArray(string.concat(".configs[", index_, "].specific.rewardTokens")));
@@ -73,25 +64,21 @@ contract AnyCompounderNaiveTest is AnyBaseTest {
 
         vm.prank(bob);
         strategy.deposit(testConfig.defaultAmount, bob);
-    
-        uint totalAssets = strategy.totalAssets();
-        uint totalSupply = strategy.totalSupply();
+
+        uint256 totalAssets = strategy.totalAssets();
+        uint256 totalSupply = strategy.totalSupply();
 
         // we give the strategy reward tokens to simulate a harvest
         address[] memory rewardTokens = strategy.rewardTokens();
-        for (uint i; i < rewardTokens.length; i++) {
+        for (uint256 i; i < rewardTokens.length; i++) {
             deal(rewardTokens[i], address(strategy), 1e18);
         }
-        
+
         // give this contract yield assets and allow the strategy to pull them
         _prepareConversion(yieldAsset, testConfig.defaultAmount);
 
         ClaimContract claimContract = new ClaimContract(rewardTokens);
-        strategy.harvest(abi.encode(
-            claimContract,
-            bytes(""),
-            testConfig.defaultAmount
-        ));
+        strategy.harvest(abi.encode(claimContract, bytes(""), testConfig.defaultAmount));
 
         assertGt(strategy.totalAssets(), totalAssets, "total assets should increase");
         assertEq(strategy.totalSupply(), totalSupply, "total supply should not change");
@@ -104,24 +91,19 @@ contract AnyCompounderNaiveTest is AnyBaseTest {
 
         vm.prank(bob);
         strategy.deposit(testConfig.defaultAmount, bob);
-    
+
         // we give the strategy reward tokens to simulate a harvest
         address[] memory rewardTokens = strategy.rewardTokens();
-        for (uint i; i < rewardTokens.length; i++) {
+        for (uint256 i; i < rewardTokens.length; i++) {
             deal(rewardTokens[i], address(strategy), 1e18);
         }
-        
+
         // give this contract yield assets and allow the strategy to pull them
         _prepareConversion(yieldAsset, testConfig.defaultAmount);
 
         ClaimContract claimContract = new ClaimContract(rewardTokens);
 
         vm.expectRevert(AnyCompounderNaive.HarvestFailed.selector);
-        strategy.harvest(abi.encode(
-            claimContract,
-            bytes(""),
-            0
-        ));
+        strategy.harvest(abi.encode(claimContract, bytes(""), 0));
     }
 }
-
